Moved "toFromUTM" to utils
